updating vmixing & tests
