+++ modules/correlation/entropy.cpp
@@ -8,7 +8,7 @@ Copyright (c) 2019 Panda Team
 #ifndef _METRIC_DISTANCE_K_RANDOM_ENTROPY_CPP
 #define _METRIC_DISTANCE_K_RANDOM_ENTROPY_CPP
 
-#include "../../modules/utils/type_traits.hpp"
+//#include "../../modules/utils/type_traits.hpp"
 #include "../../modules/space/tree.hpp"
 #include "estimator_helpers.hpp"
 #include "epmgp.hpp"
@@ -638,6 +638,172 @@ double Entropy<RecType, Metric>::estimate(
 {
     return entropy_details::estimate(a, *this, sampleSize, threshold, maxIterations);
 }
+
+
+
+
+// --------------------------- VOI & VMixing
+
+
+namespace voi_details {
+
+template <typename C1, typename C2, typename T=type_traits::underlying_type_t<C1>>
+std::vector<std::vector<T>> combine(const C1& X, const C2& Y)
+{
+    std::size_t N = X.size();
+    std::size_t dx = X[0].size();
+    std::size_t dy = Y[0].size();
+    std::vector<std::vector<T>> XY(N);
+    for (std::size_t i = 0; i < N; i++) {
+        XY[i].resize(dx + dy);
+        std::size_t k = 0;
+        for (std::size_t j = 0; j < dx; j++, k++) {
+            XY[i][k] = X[i][j];
+        }
+        for (std::size_t j = 0; j < dy; j++, k++) {
+            XY[i][k] = Y[i][j];
+        }
+    }
+    return XY;
+}
+
+
+}
+
+
+
+
+template <typename C, typename Metric>
+typename std::enable_if_t<!type_traits::is_container_of_integrals_v<C>, type_traits::underlying_type_t<C>>
+variationOfInformationSimple(const C& Xc, const C& Yc, int k)
+{
+    using T = type_traits::underlying_type_t<C>;
+
+    auto N = Xc.size();
+
+    if (N < k + 1 || Yc.size() < k + 1)
+        throw std::invalid_argument("number of points in dataset must be larger than k");
+
+    std::vector<std::vector<T>> X;
+    for (const auto& e: Xc)
+        X.push_back(std::vector<T>(std::begin(e), std::end(e))); // TODO optimize
+
+    std::vector<std::vector<T>> Y;
+    for (const auto& e: Yc)
+        Y.push_back(std::vector<T>(std::begin(e), std::end(e)));
+
+    std::vector<std::vector<T>> XY = voi_details::combine(X, Y);
+
+    auto e = EntropySimple<void, Metric>(Metric(), k);
+
+    //auto eX = e(Xc); // for debug, TODO remove
+    //auto eY = e(Yc);
+    //auto eXY = e(XY);
+    //std::cout << "    " << eX << " | " << eY << " | " << eXY << "\n";
+
+    auto result = 2 * e(XY) - e(Xc) - e(Yc);
+    //auto result = eX + eY - 2 * eXY; // TODO remove
+    //if (result < 0)
+        //return 0;
+    return result;
+}
+
+
+template <typename C, typename Metric>
+typename std::enable_if_t<!type_traits::is_container_of_integrals_v<C>, type_traits::underlying_type_t<C>>
+variationMixedSimple(const C& Xc, const C& Yc, int k)
+{
+    using T = type_traits::underlying_type_t<C>;
+
+    auto N = Xc.size();
+
+    if (N < k + 1 || Yc.size() < k + 1)
+        throw std::invalid_argument("number of points in dataset must be larger than k");
+
+    std::vector<std::vector<T>> X;
+    for (const auto& e: Xc)
+        X.push_back(std::vector<T>(std::begin(e), std::end(e))); // TODO optimize
+
+    std::vector<std::vector<T>> Y;
+    for (const auto& e: Yc)
+        Y.push_back(std::vector<T>(std::begin(e), std::end(e)));
+
+    std::vector<std::vector<T>> XY; // concatenation instead of combine(X, Y);
+    XY.reserve(X.size() + Y.size());
+    XY.insert(XY.end(), X.begin(), X.end());
+    XY.insert(XY.end(), Y.begin(), Y.end());
+
+    auto e = EntropySimple<void, Metric>(Metric(), k);
+    auto result = 2 * e(XY) - e(Xc) - e(Yc);
+    return result;
+}
+
+
+template <typename C, typename Metric>
+typename std::enable_if_t<!type_traits::is_container_of_integrals_v<C>, type_traits::underlying_type_t<C>>
+variationOfInformation_kpN(const C& Xc, const C& Yc, int k, int p)
+{
+    using T = type_traits::underlying_type_t<C>;
+
+    auto N = Xc.size();
+
+    if (N < k + 1 || Yc.size() < k + 1)
+        throw std::invalid_argument("number of points in dataset must be larger than k");
+
+    std::vector<std::vector<T>> X;
+    for (const auto& e: Xc)
+        X.push_back(std::vector<T>(std::begin(e), std::end(e))); // TODO optimize
+
+    std::vector<std::vector<T>> Y;
+    for (const auto& e: Yc)
+        Y.push_back(std::vector<T>(std::begin(e), std::end(e)));
+
+    std::vector<std::vector<T>> XY = voi_details::combine(X, Y);
+
+    auto e = Entropy<void, Metric>(Metric(), k, p);
+
+    //auto eX = e(Xc); // for debug, TODO remove
+    //auto eY = e(Yc);
+    //auto eXY = e(XY);
+    //std::cout << "    " << eX << " | " << eY << " | " << eXY << "\n";
+
+    auto result = 2 * e(XY) - e(Xc) - e(Yc);
+    //auto result = eX + eY - 2 * eXY; // TODO remove
+    //if (result < 0)
+        //return 0;
+    return result;
+}
+
+
+template <typename C, typename Metric>
+typename std::enable_if_t<!type_traits::is_container_of_integrals_v<C>, type_traits::underlying_type_t<C>>
+variationMixed_kpN(const C& Xc, const C& Yc, int k, int p)
+{
+    using T = type_traits::underlying_type_t<C>;
+
+    auto N = Xc.size();
+
+    if (N < k + 1 || Yc.size() < k + 1)
+        throw std::invalid_argument("number of points in dataset must be larger than k");
+
+    std::vector<std::vector<T>> X;
+    for (const auto& e: Xc)
+        X.push_back(std::vector<T>(std::begin(e), std::end(e))); // TODO optimize
+
+    std::vector<std::vector<T>> Y;
+    for (const auto& e: Yc)
+        Y.push_back(std::vector<T>(std::begin(e), std::end(e)));
+
+    std::vector<std::vector<T>> XY; // concatenation instead of combine(X, Y);
+    XY.reserve(X.size() + Y.size());
+    XY.insert(XY.end(), X.begin(), X.end());
+    XY.insert(XY.end(), Y.begin(), Y.end());
+
+    auto e = Entropy<void, Metric>(Metric(), k, p);
+    auto result = 2 * e(XY) - e(Xc) - e(Yc);
+    return result;
+}
+
