+++ modules/correlation/entropy.hpp
@@ -11,7 +11,7 @@ Copyright (c) 2020 Panda Team
 
 #include "../../3rdparty/blaze/Blaze.h"
 #include "../distance/k-related/Standards.hpp"
-
+#include "../modules/utils/type_traits.hpp"
 
 namespace metric {
 
@@ -80,6 +80,70 @@ private:
     Metric metric;
     bool exp;
 };
+
+
+// VOI & VMixing
+
+
+/**
+ * @brief
+ *
+ * @param Xc
+ * @param Yc
+ * @param k
+ * @param logbase
+ * @return
+ */
+template <typename C, typename Metric = metric::Chebyshev<type_traits::underlying_type_t<C>>>
+typename std::enable_if_t<!type_traits::is_container_of_integrals_v<C>, type_traits::underlying_type_t<C>>
+variationOfInformationSimple(const C& Xc, const C& Yc, int k = 3);
+
+
+/**
+ * @brief
+ *
+ * @param Xc
+ * @param Yc
+ * @param k
+ * @param logbase
+ * @return
+ */
+template <typename C, typename Metric = metric::Chebyshev<type_traits::underlying_type_t<C>>>
+typename std::enable_if_t<!type_traits::is_container_of_integrals_v<C>, type_traits::underlying_type_t<C>>
+variationMixedSimple(const C& Xc, const C& Yc, int k = 3);
+
+
+
+/**
+ * @brief
+ *
+ * @param Xc
+ * @param Yc
+ * @param k
+ * @param logbase
+ * @return
+ */
+template <typename C, typename Metric = metric::Chebyshev<type_traits::underlying_type_t<C>>>
+typename std::enable_if_t<!type_traits::is_container_of_integrals_v<C>, type_traits::underlying_type_t<C>>
+variationOfInformation_kpN(const C& Xc, const C& Yc, int k = 3, int p = 25);
+
+
+
+/**
+ * @brief
+ *
+ * @param Xc
+ * @param Yc
+ * @param k
+ * @param logbase
+ * @return
+ */
+template <typename C, typename Metric = metric::Chebyshev<type_traits::underlying_type_t<C>>>
+typename std::enable_if_t<!type_traits::is_container_of_integrals_v<C>, type_traits::underlying_type_t<C>>
+variationMixed_kpN(const C& Xc, const C& Yc, int k = 3, int p = 25);
+
+
+
